ch5 practice set 5 done

+++ C5_Functions_and_Recursion/06_practice_1.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
-int average(int,int,int);
+float average(int,int,int);
 
-int average(int a,int b,int c){
-    return ((a+b+c)/3);
+float average(int a,int b,int c){
+    return ((a+b+c)/3.0);  // why 3.0 because we have to make it float 
 }
 
 int main() {
@@ -14,6 +14,6 @@ int main() {
     scanf("%d",&b);
     printf("Enter the No3:");
     scanf("%d",&c);
-    printf("The averafe of 3 nos %d , %d and %d is %d.\n",a,b,c,average(a,b,c));
+    printf("The averafe of 3 nos %d , %d and %d is %f.\n",a,b,c,average(a,b,c));
     return 0;
 }
